resizer/requirements.txt
wand
google-cloud-storage